requirements.txt
flask==2.2.5
werkzeug==2.2.3
kafka-python==2.2.15
faust-streaming==0.11.3
requests==2.32.5
python-dotenv==1.0.1
pydantic==2.8.0
Flask-SQLAlchemy==3.0.5
cryptography==41.0.3
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.8.0
incdbscan==0.4.0
matplotlib==3.9.0
folium==0.17.0
